Make iter map more closely to rocksdb, tests
Add tests for iterator and set/get

Change to iterator interface to more closely mirror rocksdb.

If want to just have a `next()` zig-style, would need to either:
- have a check on the "first" entry, since you would either seek to
  the first entry or advance to the next.
- if you want to not have a branch at the beginning of `next`, need to
  advance to next entry at the end of `next`, but then you need to
  allocate the entry kv, since rocksdb is using ptrs to internal buffer.

So the current solution to mirror the rocksdb interface means a bit more
manual advancement of the iterator, but fewer branches/allocations.